refactor: move `from/to_ascii` outside of vial
